requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
